chore: delete unused files and rename NewKeyService to KeyService
