+++ jans-cedarling/cedarling/src/jwt/key_service.rs
@@ -4,33 +4,52 @@
 // Copyright (c) 2024, Gluu, Inc.
 
 use std::collections::HashMap;
-use std::sync::Arc;
-use std::time::Duration;
 
-use jsonwebtoken::DecodingKey;
-use serde_json::{Value, json};
+use jsonwebtoken::jwk::{Jwk, JwkSet, KeyAlgorithm};
+use jsonwebtoken::{Algorithm, DecodingKey};
+use reqwest::Client;
+use serde::{Deserialize, Deserializer, de};
+use url::Url;
 
-use super::TrustedIssuerId;
-use super::jwk_store::{JwkStore, JwkStoreError};
 use crate::common::policy_store::TrustedIssuer;
-use crate::http::{HttpClient, HttpClientError};
 
-pub struct DecodingKeyWithIss<'a> {
-    /// The decoding key used to validate JWT signatures.
-    pub key: &'a DecodingKey,
-    /// The Trusted Issuer where the Key was fetched.
-    pub key_iss: Option<&'a TrustedIssuer>,
+#[derive(Debug, Hash, Eq, PartialEq)]
+pub struct DecodingKeyInfo {
+    pub issuer: Option<String>,
+    pub kid: Option<String>,
+    pub algorithm: Algorithm,
 }
 
-/// Manages Json Web Keys (JWK).
-// TODO: periodically update the key stores to ensure keys are valid.
-#[derive(Debug)]
-#[deprecated]
+#[derive(Deserialize)]
+struct OpenIdConfig {
+    issuer: String,
+    #[serde(deserialize_with = "deserialize_url")]
+    jwks_uri: Url,
+}
+
+pub fn deserialize_url<'de, D>(deserializer: D) -> Result<Url, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let url_str = String::deserialize(deserializer)?;
+    let url = Url::parse(&url_str)
+        .map_err(|e| de::Error::custom(format!("invalid url '{url_str}': {e}")))?;
+    Ok(url)
+}
+
+/// Manages JSON Web Keys (JWKs) used for decoding JWTs.
+// TODO: some IDPs might rotate their keys so we need to figure out a way to keep
+// the keys updated.
+#[derive(Default)]
 pub struct KeyService {
-    key_stores: HashMap<TrustedIssuerId, JwkStore>,
+    keys: HashMap<DecodingKeyInfo, DecodingKey>,
 }
 
 impl KeyService {
+    pub fn new() -> Self {
+        Default::default()
+    }
+
     /// Loads JWK stores from a string.
     ///
     /// This enables loading keystores via a local JSON file.
@@ -50,95 +69,159 @@ impl KeyService {
     /// - and the values contains the JSON Web Keys as defined in [`RFC 7517`].
     ///
     /// [`RFC 7517`]: https://datatracker.ietf.org/doc/html/rfc7517
-    pub fn new_from_str(key_stores: &str) -> Result<Self, KeyServiceError> {
-        let parsed_stores = serde_json::from_str::<HashMap<String, Value>>(key_stores)
-            .map_err(KeyServiceError::DecodeJwkStores)?;
-        let mut key_stores = HashMap::new();
-        for (iss_id, keys) in &parsed_stores {
-            let iss_id = TrustedIssuerId::from(iss_id.as_str());
-            let jwks = json!({"keys": keys});
-            key_stores.insert(iss_id.clone(), JwkStore::new_from_jwks_value(iss_id, jwks)?);
-        }
-        Ok(Self { key_stores })
-    }
+    pub fn insert_keys_from_str(&mut self, key_stores: &str) -> Result<(), KeyServiceError> {
+        let parsed_stores = serde_json::from_str::<HashMap<String, Vec<Jwk>>>(key_stores)
+            .map_err(InsertKeysError::DeserializeJwkStores)?;
 
-    /// Loads key stores using a JSON string.
-    ///
-    /// Enables loading key stores from a local JSON file.
-    pub async fn new_from_trusted_issuers(
-        trusted_issuers: &HashMap<String, TrustedIssuer>,
-    ) -> Result<Self, KeyServiceError> {
-        let http_client = HttpClient::new(3, Duration::from_secs(3))?;
-
-        let mut key_stores = HashMap::new();
-        for (iss_id, iss) in trusted_issuers.iter() {
-            let iss_id: Arc<str> = iss_id.as_str().into();
-            key_stores.insert(
-                iss_id.clone(),
-                JwkStore::new_from_trusted_issuer(iss_id, iss, &http_client).await?,
-            );
-        }
-
-        Ok(Self { key_stores })
-    }
-
-    /// Gets the decoding key with the given key ID from the store with it's Trusted Issuer.
-    pub fn get_key(&self, key_id: &str) -> Option<DecodingKeyWithIss> {
-        // PERF: We can add a reference of all the keys into a HashMap
-        // so we do not need to loop through all of these.
-        for store in self.key_stores.values() {
-            if let Some(key) = store.get(key_id) {
-                return Some(DecodingKeyWithIss {
-                    key,
-                    key_iss: store.source_iss(),
-                });
+        for (issuer, keys) in parsed_stores.into_iter() {
+            for jwk in keys.into_iter() {
+                let decoding_key =
+                    DecodingKey::from_jwk(&jwk).map_err(InsertKeysError::BuildDecodingKey)?;
+                let algorithm = get_algorithm(&jwk)
+                    .map_err(InsertKeysError::UnsupportedKeyAlgorithm)?
+                    .ok_or(InsertKeysError::UnspecifiedAlgorithm)?;
+                let key_info = DecodingKeyInfo {
+                    issuer: Some(issuer.clone()),
+                    kid: jwk.common.key_id,
+                    algorithm,
+                };
+                self.keys.insert(key_info, decoding_key);
             }
         }
 
-        None
+        Ok(())
     }
 
-    /// Returns a Vec containing a reference to all of the keys.
-    ///
-    /// Useful if the keys from the JWKS do not have a `kid` (Key ID).
-    //  this is currently unused since we don't handle JWTs without `kid` yet.
-    #[allow(dead_code)]
-    pub fn get_keys(&self) -> Vec<&DecodingKey> {
-        // PERF: We can cache the returned Vec so it doesn't
-        // get created every time this function is called.
-        let mut keys = Vec::new();
-        self.key_stores
-            .values()
-            .for_each(|store| keys.extend(store.get_keys()));
-        keys
+    pub async fn fetch_keys_for_iss(&mut self, iss: &TrustedIssuer) -> Result<(), KeyServiceError> {
+        let client = Client::new();
+
+        let openid_config = client
+            .get(iss.oidc_endpoint.as_str())
+            .send()
+            .await
+            .map_err(FetchKeysError::GetOpenIdConfig)?
+            .error_for_status()
+            .map_err(FetchKeysError::GetOpenIdConfig)?
+            .json::<OpenIdConfig>()
+            .await
+            .map_err(FetchKeysError::DeserializeOpenIdConfig)?;
+
+        let jwks = client
+            .get(openid_config.jwks_uri)
+            .send()
+            .await
+            .map_err(FetchKeysError::GetJwks)?
+            .error_for_status()
+            .map_err(FetchKeysError::GetJwks)?
+            .json::<JwkSet>()
+            .await
+            .map_err(FetchKeysError::DeserializeJwks)?;
+
+        for jwk in jwks.keys.into_iter() {
+            let key = DecodingKey::from_jwk(&jwk).map_err(FetchKeysError::BuildDecodingKey)?;
+            let algorithm = get_algorithm(&jwk)
+                .map_err(FetchKeysError::UnsupportedKeyAlgorithm)?
+                .ok_or(FetchKeysError::UnspecifiedAlgorithm)?;
+            let key_info = DecodingKeyInfo {
+                issuer: Some(openid_config.issuer.clone()),
+                kid: jwk.common.key_id,
+                algorithm,
+            };
+            self.keys.insert(key_info, key);
+        }
+
+        Ok(())
     }
+
+    pub fn get_key(&self, key_info: &DecodingKeyInfo) -> Option<&DecodingKey> {
+        self.keys.get(key_info)
+    }
+
+    pub fn has_keys(&self) -> bool {
+        !self.keys.is_empty()
+    }
+}
+
+#[inline]
+fn get_algorithm(jwk: &Jwk) -> Result<Option<Algorithm>, KeyAlgorithm> {
+    let Some(algorithm) = jwk.common.key_algorithm else {
+        return Ok(None);
+    };
+
+    let algorithm = match algorithm {
+        KeyAlgorithm::HS256 => Algorithm::HS256,
+        KeyAlgorithm::HS384 => Algorithm::HS384,
+        KeyAlgorithm::HS512 => Algorithm::HS512,
+        KeyAlgorithm::ES256 => Algorithm::ES256,
+        KeyAlgorithm::ES384 => Algorithm::ES384,
+        KeyAlgorithm::RS256 => Algorithm::RS256,
+        KeyAlgorithm::RS384 => Algorithm::RS384,
+        KeyAlgorithm::RS512 => Algorithm::RS512,
+        KeyAlgorithm::PS256 => Algorithm::PS256,
+        KeyAlgorithm::PS384 => Algorithm::PS384,
+        KeyAlgorithm::PS512 => Algorithm::PS512,
+        KeyAlgorithm::EdDSA => Algorithm::EdDSA,
+        alg => return Err(alg),
+    };
+
+    Ok(Some(algorithm))
 }
 
 #[derive(thiserror::Error, Debug)]
 pub enum KeyServiceError {
-    #[error("Failed to decode JWK Stores from string: {0}")]
-    DecodeJwkStores(serde_json::Error),
-    #[error("Failed to make HTTP Request: {0}")]
-    Http(#[from] HttpClientError),
-    #[error("Failed to load JWKS: {0}")]
-    JwkStoreError(#[from] JwkStoreError),
+    #[error("failed to insert keys into the KeyService: {0}")]
+    InsertKeys(#[from] InsertKeysError),
+    #[error("failed to fetch keys for the KeyService: {0}")]
+    FetchKeysError(#[from] FetchKeysError),
+}
+
+/// Errors encountered while inserting keys using strings
+#[derive(thiserror::Error, Debug)]
+pub enum InsertKeysError {
+    #[error("failed to deserialize string into JWK stores: {0}")]
+    DeserializeJwkStores(#[from] serde_json::Error),
+    #[error("unsupported key algorithm: {0}")]
+    UnsupportedKeyAlgorithm(KeyAlgorithm),
+    #[error("failed to build decoding key: {0}")]
+    BuildDecodingKey(#[from] jsonwebtoken::errors::Error),
+    #[error("the key did not specify it's algorithm")]
+    UnspecifiedAlgorithm,
+}
+
+/// Errors encountered while fetching keys remotely
+#[derive(thiserror::Error, Debug)]
+pub enum FetchKeysError {
+    #[error("failed to get openid config: {0}")]
+    GetOpenIdConfig(#[source] reqwest::Error),
+    #[error("failed to deserialize openid config: {0}")]
+    DeserializeOpenIdConfig(#[source] reqwest::Error),
+    #[error("failed to get JWKS: {0}")]
+    GetJwks(#[source] reqwest::Error),
+    #[error("failed to deserialize JWKS: {0}")]
+    DeserializeJwks(#[source] reqwest::Error),
+    #[error("unsupported key algorithm: {0}")]
+    UnsupportedKeyAlgorithm(KeyAlgorithm),
+    #[error("failed to build decoding key: {0}")]
+    BuildDecodingKey(#[from] jsonwebtoken::errors::Error),
+    #[error("the key did not specify it's algorithm")]
+    UnspecifiedAlgorithm,
 }
 
 #[cfg(test)]
 mod test {
-    use crate::common::policy_store::TrustedIssuer;
-    use crate::jwt::key_service::KeyService;
-    use mockito::Server;
+    use super::*;
+    use crate::jwt::test_utils::MockServer;
+    use jsonwebtoken::Algorithm;
     use serde_json::json;
-    use std::collections::HashMap;
-    use url::Url;
 
     #[test]
-    fn can_load_from_str() {
-        let kid1 = "a50f6e70ef4b548a5fd9142eecd1fb8f54dce9ee";
-        let kid2 = "73e25f9789119c7875d58087a78ac23f5ef2eda3";
+    fn can_insert_and_get_keys_from_str() {
+        let iss1 = "http://some_issuer.com".to_string();
+        let kid1 = "a50f6e70ef4b548a5fd9142eecd1fb8f54dce9ee".to_string();
+        let iss2 = "http://another_issuer.com".to_string();
+        let kid2 = "73e25f9789119c7875d58087a78ac23f5ef2eda3".to_string();
         let key_stores = json!({
-            "some_issuer_id": [{
+            &iss1: [{
                 "use": "sig",
                 "e": "AQAB",
                 "alg": "RS256",
@@ -146,7 +229,7 @@ mod test {
                 "kty": "RSA",
                 "kid": "a50f6e70ef4b548a5fd9142eecd1fb8f54dce9ee"
             }],
-            "anorther_issuer_id": [{
+            &iss2: [{
                 "n": "tMXbmw7xEDVLLkAJdxpI-6pGywn0x9fHbD_mfgtFGZEs1LDjhDAJq6c-SoODeWQstjpetTgNqVCKOuU6zGyFPNtkDjhJqDW6THy06uJ8I85crILo3h-6NPclZ3bK9OzN5bIbzjbSvxrIM7ORZOlWzByOn5qGsMvI3aDrZ0lXNC1eCDWJpoJznG1fWcHYxbUy_CHDC3Cd26jX19aRALEEQU-y-wi9pv86qxEmrYMLsVN3__eWNNPkzxgf0eSOWFDv5_19YK7irYztqiwin6abxr9RHj3Qs21hpJ9A-YfsfmNkxmifgDeiTnXpZY8yfVTCJTtkgT7sjdU1lvhsMa4Z0w",
                 "e": "AQAB",
                 "use": "sig",
@@ -155,145 +238,120 @@ mod test {
                 "kid": "73e25f9789119c7875d58087a78ac23f5ef2eda3"
             }],
         });
+        let mut key_service = KeyService::default();
 
-        let key_service = KeyService::new_from_str(&key_stores.to_string())
-            .expect("Should load KeyService from str");
+        key_service
+            .insert_keys_from_str(&key_stores.to_string())
+            .expect("insert keys");
 
         assert!(
-            key_service.get_key(kid1).is_some(),
+            key_service
+                .get_key(&DecodingKeyInfo {
+                    issuer: Some(iss1),
+                    kid: Some(kid1.clone()),
+                    algorithm: Algorithm::RS256,
+                })
+                .is_some(),
             "Expected to find a key"
         );
+
         assert!(
-            key_service.get_key(kid2).is_some(),
+            key_service
+                .get_key(&DecodingKeyInfo {
+                    issuer: Some(iss2),
+                    kid: Some(kid2.clone()),
+                    algorithm: Algorithm::RS256,
+                })
+                .is_some(),
             "Expected to find a key"
         );
+
         assert!(
-            key_service.get_key("some unknown key id").is_none(),
+            key_service
+                .get_key(&DecodingKeyInfo {
+                    issuer: Some("some_unknown_iss".to_string()),
+                    kid: Some(kid1),
+                    algorithm: Algorithm::RS256,
+                })
+                .is_none(),
+            "Expected to not find a key"
+        );
+
+        assert!(
+            key_service
+                .get_key(&DecodingKeyInfo {
+                    issuer: None,
+                    kid: Some(kid2),
+                    algorithm: Algorithm::HS256,
+                })
+                .is_none(),
             "Expected to not find a key"
         );
     }
 
     #[tokio::test]
     async fn can_load_jwk_stores_from_multiple_trusted_issuers() {
-        let kid1 = "a50f6e70ef4b548a5fd9142eecd1fb8f54dce9ee";
-        let kid2 = "73e25f9789119c7875d58087a78ac23f5ef2eda3";
+        let server1 = MockServer::new_with_defaults().await.unwrap();
+        let iss1 = server1.trusted_issuer();
+        let (_key1, kid1) = server1.jwt_decoding_key_and_id().unwrap();
 
-        let mut mock_server = Server::new_async().await;
+        let server2 = MockServer::new_with_defaults().await.unwrap();
+        let iss2 = server2.trusted_issuer();
+        let (_key2, kid2) = server2.jwt_decoding_key_and_id().unwrap();
 
-        // Setup first OpenID config endpoint
-        let openid_config_endpoint1 = mock_server
-            .mock("GET", "/first/.well-known/openid-configuration")
-            .with_status(200)
-            .with_body(
-                json!({
-                    "issuer": mock_server.url(),
-                    "jwks_uri": format!("{}/first/jwks", mock_server.url())
+        let mut key_service = KeyService::default();
+
+        key_service
+            .fetch_keys_for_iss(&iss1)
+            .await
+            .expect("fetch keys for issuer 1");
+        key_service
+            .fetch_keys_for_iss(&iss2)
+            .await
+            .expect("fetch keys for issuer 2");
+
+        assert!(
+            key_service
+                .get_key(&DecodingKeyInfo {
+                    issuer: Some(server1.issuer()),
+                    kid: kid1,
+                    algorithm: Algorithm::HS256,
                 })
-                .to_string(),
-            )
-            .expect(1)
-            .create();
-        // Setup first JWKS endpoint
-        let jwks_endpoint1 = mock_server
-            .mock("GET", "/first/jwks")
-            .with_status(200)
-            .with_body(
-                json!({
-                "keys": [
-                    {
-                        "use": "sig",
-                        "e": "AQAB",
-                        "alg": "RS256",
-                        "n": "4VI56fF0rcWHHVgHFLHrmEO5w8oN9gbSQ9TEQnlIKRg0zCtl2dLKtt0hC6WMrTA9cF7fnK4CLNkfV_Mytk-rydu2qRV_kah62v9uZmpbS5dcz5OMXmPuQdV8fDVIvscDK5dzkwD3_XJ2mzupvQN2reiYgce6-is23vwOyuT-n4vlxSqR7dWdssK5sj9mhPBEIlfbuKNykX5W6Rgu-DyuoKArc_aukWnLxWN-yoroP2IHYdCQm7Ol08vAXmrwMyDfvsmqdXUEx4om1UZ5WLf-JNaZp4lXhgF7Cur5066213jwpp4f_D3MyR-oa43fSa91gqp2berUgUyOWdYSIshABVQ",
-                        "kty": "RSA",
-                        "kid": kid1,
-                    },
-                ]
-                }).to_string())
-            .expect(1)
-            .create();
+                .is_some(),
+            "expected to find a key from issuer 1"
+        );
 
-        // Setup second OpenID config endpoint
-        let openid_config_endpoint2 = mock_server
-            .mock("GET", "/second/.well-known/openid-configuration")
-            .with_status(200)
-            .with_body(
-                json!({
-                    "issuer": mock_server.url(),
-                    "jwks_uri": format!("{}/second/jwks", mock_server.url())
+        assert!(
+            key_service
+                .get_key(&DecodingKeyInfo {
+                    issuer: Some(server2.issuer()),
+                    kid: kid2.clone(),
+                    algorithm: Algorithm::HS256,
                 })
-                .to_string(),
-            )
-            .expect(1)
-            .create();
-        // Setup second JWKS endpoint
-        let jwks_endpoint2 = mock_server
-            .mock("GET", "/second/jwks")
-            .with_status(200)
-            .with_body(
-                json!({
-                "keys": [
-                    {
-                        "n": "tMXbmw7xEDVLLkAJdxpI-6pGywn0x9fHbD_mfgtFGZEs1LDjhDAJq6c-SoODeWQstjpetTgNqVCKOuU6zGyFPNtkDjhJqDW6THy06uJ8I85crILo3h-6NPclZ3bK9OzN5bIbzjbSvxrIM7ORZOlWzByOn5qGsMvI3aDrZ0lXNC1eCDWJpoJznG1fWcHYxbUy_CHDC3Cd26jX19aRALEEQU-y-wi9pv86qxEmrYMLsVN3__eWNNPkzxgf0eSOWFDv5_19YK7irYztqiwin6abxr9RHj3Qs21hpJ9A-YfsfmNkxmifgDeiTnXpZY8yfVTCJTtkgT7sjdU1lvhsMa4Z0w",
-                        "e": "AQAB",
-                        "use": "sig",
-                        "alg": "RS256",
-                        "kty": "RSA",
-                        "kid": kid2,
-                    }
-
-                ]}).to_string())
-            .expect(1)
-            .create();
-
-        let key_service = KeyService::new_from_trusted_issuers(&HashMap::from([
-            (
-                "first".to_string(),
-                TrustedIssuer {
-                    name: "First IDP".to_string(),
-                    description: "".to_string(),
-                    oidc_endpoint: Url::parse(&format!(
-                        "{}/first/.well-known/openid-configuration",
-                        mock_server.url()
-                    ))
-                    .expect("should be a valid URL"),
-                    ..Default::default()
-                },
-            ),
-            (
-                "second".to_string(),
-                TrustedIssuer {
-                    name: "Second IDP".to_string(),
-                    description: "".to_string(),
-                    oidc_endpoint: Url::parse(&format!(
-                        "{}/second/.well-known/openid-configuration",
-                        mock_server.url()
-                    ))
-                    .expect("should be a valid URL"),
-                    ..Default::default()
-                },
-            ),
-        ]))
-        .await
-        .expect("Should load KeyService from trusted issuers");
-
-        assert!(
-            key_service.get_key(kid1).is_some(),
-            "Expected to find a key"
-        );
-        assert!(
-            key_service.get_key(kid2).is_some(),
-            "Expected to find a key"
-        );
-        assert!(
-            key_service.get_key("some unknown key id").is_none(),
-            "Expected to not find a key"
+                .is_some(),
+            "expected to find a key from issuer 2"
         );
 
-        // Assert that each of the endpoints are only visited once.
-        openid_config_endpoint1.assert();
-        openid_config_endpoint2.assert();
-        jwks_endpoint1.assert();
-        jwks_endpoint2.assert();
+        assert!(
+            key_service
+                .get_key(&DecodingKeyInfo {
+                    issuer: Some(server2.issuer()),
+                    kid: kid2,
+                    algorithm: Algorithm::RS256,
+                })
+                .is_none(),
+            "expected not to find a key with a different algorithm"
+        );
+
+        assert!(
+            key_service
+                .get_key(&DecodingKeyInfo {
+                    issuer: Some("http://some_unknown_issuer.com".into()),
+                    kid: None,
+                    algorithm: Algorithm::HS256,
+                })
+                .is_none(),
+            "expected to not find a key from an unknown issuer"
+        );
     }
 }
